+++ src/api/products/ledgerscan/ledgerscan.controller.ts
@@ -1,7 +1,7 @@
 import { BadRequestException, Controller, Post, Req } from "@nestjs/common"
 import { LedgerscanService } from "./ledgerscan.service"
 import { envConfig } from "src/config/env.config"
-import { ApiKeyAuthorizer, ApiKeyAuthorizerResponse } from "src/authorization/apikey-authorizer.decorator"
+import { CredentialAuthorizer, CredentialAuthorizerResponse } from "src/authorization/credential-authorizer.decorator"
 import { statusMessages } from "src/constants/status-messages"
 
 @Controller("products/ledgerscan")
@@ -9,11 +9,11 @@ export class LedgerscanController {
   constructor(private readonly ledgerscanService: LedgerscanService) { }
 
   @Post("analyzer")
-  async analyze(@ApiKeyAuthorizer() ufak: ApiKeyAuthorizerResponse, @Req() req: any) {
+  async analyze(@CredentialAuthorizer() ufc: CredentialAuthorizerResponse, @Req() req: any) {
     try {
       const queryParams = req.url.split("?")[1]
       const queryParamsWithSecretKey = `${queryParams}&apiKey=${envConfig.polygonscanSecretKey}`
-      const response = await this.ledgerscanService.analyze(queryParamsWithSecretKey, ufak.workspaceId)
+      const response = await this.ledgerscanService.analyze(queryParamsWithSecretKey, ufc.workspaceId)
       return response
     }