+++ src/api/subscription/subscription.service.ts
@@ -25,9 +25,10 @@ export class SubscriptionService {
       const user = await this.userRepository.findUserById(userId)
 
       if (user.trialAvailable) {
+        const clientId = randomBytes(16).toString("hex")
+        const clientSecret = randomBytes(32).toString("hex")
         const selectedPlan = "Trial"
-        const apiKey = "ak-" + randomBytes(16).toString("hex")
-        await this.subscriptionRepository.createNewSubscription(workspaceId, selectedPlan, apiKey)
+        await this.subscriptionRepository.createNewSubscription(workspaceId, selectedPlan, clientId, clientSecret)
         await this.userRepository.findUserByIdAndUpdateTrialStatus(userId, false)
         return true
       }
@@ -62,16 +63,19 @@ export class SubscriptionService {
 
         else {
           const subscription = await this.subscriptionRepository.findSubscriptionByWorkspaceIdAndDelete(workspaceId)
-          let apiKey: string = ""
+          let clientId: string = ""
+          let clientSecret: string = ""
           if (subscription) {
-            apiKey = subscription.apiKey
+            clientId = subscription.clientId
+            clientSecret = subscription.clientSecret
           }
 
           else {
-            apiKey = "ak-" + randomBytes(16).toString("hex")
+            clientId = randomBytes(16).toString("hex")
+            clientSecret = randomBytes(32).toString("hex")
           }
 
-          await this.subscriptionRepository.createNewSubscription(workspaceId, selectedPlan, apiKey)
+          await this.subscriptionRepository.createNewSubscription(workspaceId, selectedPlan, clientId, clientSecret)
           return true
         }
       }